Added the availableMobileNumbers endpoint

+++ src/main/java/com/connection/emobile/repository/MobileNumberRepository.java
@@ -1,11 +1,10 @@
 package com.connection.emobile.repository;
 
+import com.connection.emobile.entity.MobileNumber;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.connection.emobile.entity.MobileNumber;
-
 @Repository
-public interface MobileNumberRepository extends JpaRepository<MobileNumber, Integer> {
+public interface MobileNumberRepository extends JpaRepository<MobileNumber, Long> {
 
 }
